+++ R/VarImp.R
@@ -21,93 +21,99 @@
 #'
 #' @examples
 #' data(body_fat)
-#' y=body_fat[,1]
-#' X=body_fat[,-1]
+#' y <- body_fat[, 1]
+#' X <- body_fat[, -1]
 #'
 #' tree <- ODT(X, y, split = "mse")
-#' (varimp <- VarImp(tree, type="impurity"))
-#'
-#' forest <- ODRF(X, y, split = "mse", parallel = FALSE, ntrees=50)
-#' (varimp <- VarImp(forest, type="impurity"))
-#' (varimp <- VarImp(forest, X, y, type="permutation"))
-#'
+#' (varimp <- VarImp(tree, type = "impurity"))
+#' \donttest{
+#' forest <- ODRF(X, y, split = "mse", parallel = FALSE, ntrees = 50)
+#' (varimp <- VarImp(forest, type = "impurity"))
+#' (varimp <- VarImp(forest, X, y, type = "permutation"))
+#'}
 #' @keywords forest tree
 #' @export
-VarImp <- function(obj, X=NULL, y=NULL, type="permutation") {
+VarImp <- function(obj, X = NULL, y = NULL, type = "permutation") {
+  p <- obj$data$p
+  varName <- obj$data$varName
+  pp <- obj$data$p
+  Xcat <- obj$data$Xcat
+  catLabel <- obj$data$catLabel
+  Rep <- rep(seq(p), rep(1, pp))
+  sp <- seq(pp)
+  if (!is.null(catLabel) && (sum(Xcat) > 0)) {
+    pp <- pp - length(unlist(catLabel)) + length(Xcat)
+    varName <- c(names(catLabel), varName[-seq(length(unlist(catLabel)))])
+    Rep <- rep(seq(pp), c(unlist(lapply(catLabel, length)), rep(1, pp - length(Xcat))))
+    sp <- seq(pp)
+    sp <- c(Xcat, sp[-Xcat])
+  }
 
-  p=obj[["data"]][["p"]]
+  if (type == "impurity") {
+    VarImp.impurity <- function(structure) {
+      DecImpurity <- rep(0, p)
 
-  if(type=="impurity"){
-    VarImp.impurity <- function(structure){
-      DecImpurity=rep(0,p)
-
-      nodeRotaMat=structure[["nodeRotaMat"]]
-      if(length(structure[["nodeCutIndex"]])==1){
+      nodeRotaMat <- structure[["nodeRotaMat"]]
+      if (length(structure[["nodeCutIndex"]]) == 1) {
         stop("No tree structure to measure the importance of variables!")
-      }else{
-        cutNodes=unique(nodeRotaMat[nodeRotaMat[,1]!=0,2])
+      } else {
+        cutNodes <- unique(nodeRotaMat[nodeRotaMat[, 1] != 0, 2])
       }
 
-      for (node in  cutNodes) {
-        idx=which(nodeRotaMat[,2]==node)
-        DecImpurity[nodeRotaMat[idx,1]]=DecImpurity[nodeRotaMat[idx,1]]+
-          (nodeRotaMat[idx,3])^2*structure[["nodeCutIndex"]][node]
+      for (node in cutNodes) {
+        idx <- which(nodeRotaMat[, 2] == node)
+        DecImpurity[nodeRotaMat[idx, 1]] <- DecImpurity[nodeRotaMat[idx, 1]] +
+          (nodeRotaMat[idx, 3])^2 * structure[["nodeCutIndex"]][node]
       }
 
-      DecImpurity <- DecImpurity/length(cutNodes)
+      DecImpurity <- DecImpurity / length(cutNodes)
 
       return(DecImpurity)
     }
 
-    if("ODT"%in%class(obj)){
-      DecImpurity=VarImp.impurity(obj$structure)
+    if ("ODT" %in% class(obj)) {
+      DecImpurity <- VarImp.impurity(obj$structure)
     }
 
-    if("ODRF"%in%class(obj)){
-      DecImpurity <- vapply(obj$structure, function(structure){
-        if(length(structure[["nodeCutIndex"]])==1){
-          rep(0,p)
-        }else{
+    if ("ODRF" %in% class(obj)) {
+      DecImpurity <- vapply(obj$structure, function(structure) {
+        if (length(structure[["nodeCutIndex"]]) == 1) {
+          rep(0, p)
+        } else {
           VarImp.impurity(structure)
         }
-        } , rep(0,p))
+      }, rep(0, p))
 
       DecImpurity <- rowMeans(DecImpurity)
     }
 
-    varimport <- cbind(varible = seq(p), decreas_impurity = DecImpurity)
-    rownames(varimport) <- obj[["data"]][["varName"]]
+    DecImpurity <- aggregate(DecImpurity, list(Rep), mean)[-1]
+    varimport <- cbind(sp, DecImpurity)
+    colnames(varimport) <- c("varible", "decrease_accuracy")
+    rownames(varimport) <- varName
   }
 
 
   #####################################################################
-  if(type=="permutation"){
+  if (type == "permutation") {
+    X <- as.matrix(X)
+    colnames(X) <- varName
+    if (ncol(X) != pp) {
+      stop("The dimensions of 'Xnew' and training data do not match")
+    }
 
-    if("ODT"%in%class(obj)){
+    if ("ODT" %in% class(obj)) {
       stop("Tree structure can't use permutation method to measure the importance of variables!")
     }
 
-    if("ODRF"%in%class(obj)){
-      if(is.null(X)&&is.null(y)){
+    if ("ODRF" %in% class(obj)) {
+      if (is.null(X) && is.null(y)) {
         stop("training data 'X' and 'y' argument is required")
       }
 
       if (!obj$forest$storeOOB) {
         stop("out-of-bag indices for each tree are not stored, so can't use permutation method!")
       }
-
-
-      X <- as.matrix(X)
-      colnames(X) <- obj$data$varName
-
-      pp <- obj$data$p
-      if (!is.null(obj$data$catLabel) && (sum(obj$data$Xcat) > 0)) {
-        pp <- pp - length(unlist(obj$data$catLabel)) + length(obj$data$Xcat)
-      }
-      if (ncol(X) != pp) {
-        stop("The dimensions of 'Xnew' and training data do not match")
-      }
-
 
       if (!is.null(obj$data$subset)) {
         X <- X[obj$data$subset, ]
@@ -117,8 +123,8 @@ VarImp <- function(obj, X=NULL, y=NULL, type="permutation") {
       # weights=weights0
 
 
-      if (obj$split != "mse") {
-        #y <- factor(y, levels = obj$Levels)
+      if (obj$split %in% c("gini", "entropy")) {
+        # y <- factor(y, levels = obj$Levels)
         y <- as.character(y)
       }
       # X=obj$data$X
@@ -127,8 +133,6 @@ VarImp <- function(obj, X=NULL, y=NULL, type="permutation") {
       n <- length(y)
       p <- ncol(X)
 
-      Xcat <- obj$data$Xcat
-      catLabel <- obj$data$catLabel
       numCat <- 0
       if (sum(Xcat) > 0) {
         xj <- 1
@@ -154,8 +158,8 @@ VarImp <- function(obj, X=NULL, y=NULL, type="permutation") {
         rm(X1)
         rm(Xj)
       }
-      if (!is.numeric(X)){
-        X=apply(X, 2, as.numeric)
+      if (!is.numeric(X)) {
+        X <- apply(X, 2, as.numeric)
       }
 
       # Variable scaling.
@@ -165,8 +169,8 @@ VarImp <- function(obj, X=NULL, y=NULL, type="permutation") {
           matrix(obj$data$maxminCol, n, length(indp), byrow = T)
       }
 
-      split=obj$split
-      Levels=obj$Levels
+      split <- obj$split
+      Levels <- obj$Levels
       runOOBErr <- function(structure, ...) {
         oobErrs <- rep(0, p)
         oobIndex <- structure$oobIndex
@@ -177,38 +181,40 @@ VarImp <- function(obj, X=NULL, y=NULL, type="permutation") {
         #  e.0=mean((yi-mean(y[-oobIndex]))^2)
         # }
 
-        pred <- predictTree(structure,X0,split,Levels)$prediction
-        if (split != "mse") {
+        pred <- predictTree(structure, X0, split, Levels)$prediction
+        if (split %in% c("gini", "entropy")) {
           oobErr0 <- mean(pred != y0)
         } else {
           oobErr0 <- mean((pred - y0)^2) # /e.0
         }
 
         for (j in seq(p)) {
-          Xi=X0
+          Xi <- X0
           Xi[, j] <- Xi[sample(n0), j] #+rnorm(length(oobIndex))
-          pred <- predictTree(structure,Xi,split,Levels)$prediction
-          if (split != "mse") {
+          pred <- predictTree(structure, Xi, split, Levels)$prediction
+          if (split %in% c("gini", "entropy")) {
             oobErr <- mean(pred != y0)
           } else {
             oobErr <- mean((pred - y0)^2) # /e.0
           }
 
-          oobErrs[j] <- oobErr - oobErr0#abs()
+          oobErrs[j] <- oobErr - oobErr0 # abs()
         }
 
         return(oobErrs)
       }
 
-      IncErr <- vapply(obj$structure, runOOBErr, rep(0,p))
-      #IncErr <- sapply(obj$ppTrees, runOOBErr)
+      IncErr <- vapply(obj$structure, runOOBErr, rep(0, p))
+      # IncErr <- sapply(obj$ppTrees, runOOBErr)
       IncErr <- rowMeans(IncErr)
-      varimport <- cbind(varible = seq(p), increased_error = IncErr)
-      rownames(varimport) <- colnames(X)
+      IncErr <- aggregate(IncErr, list(Rep), mean)[-1]
+      varimport <- cbind(sp, IncErr)
+      colnames(varimport) <- c("varible", "decrease_accuracy")
+      rownames(varimport) <- varName
     }
   }
 
-  varimport <- list(varImp = varimport[order(varimport[,2], decreasing = T), ], split = obj$split)
+  varimport <- list(varImp = varimport[order(varimport[, 2], decreasing = T), ], split = obj$split)
   class(varimport) <- "VarImp"
 
   return(varimport)